Use context API to provide global state data to the TODO APP
Create context store. Defined global state and functions to provide to the entire app. Created useTodosContext custom hook to use global state data in children components.
